+++ src/main/java/com/ontotext/refine/client/RefineClient.java
@@ -1,5 +1,6 @@
 package com.ontotext.refine.client;
 
+import java.io.Closeable;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -8,12 +9,11 @@ import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.utils.URIBuilder;
 import org.apache.http.impl.client.CloseableHttpClient;
 
-
 /**
  * Represents the client which purpose is to provide the HTTP communication between the application
  * and the Refine instance.
  */
-public class RefineClient implements AutoCloseable {
+public class RefineClient implements Closeable {
 
   private final URI uri;
   private final CloseableHttpClient httpClient;
@@ -58,7 +58,7 @@ public class RefineClient implements AutoCloseable {
   }
 
   @Override
-  public void close() throws Exception {
+  public void close() throws IOException {
     httpClient.close();
   }